+++ raccoon/__init__.py
@@ -2,4 +2,4 @@
 from .dataframe import DataFrame
 from .series import Series, ViewSeries
 
-__version__ = "2.0.0"
+__version__ = "2.1.0"